[Command] Move Install & Update into dedicated files
